Se arreglo la pestaña de farmacologia (Hay que solucionar para que se abara su popup)
